refactor: newChat modal and move some files
