+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { OneLiner } from "./OneLiner";
 import { Routes, Route } from "react-router-dom";
-import { ToastContainer } from "./components/ToastContainer";
+import { ToastContainer } from "./components/Common/ToastContainer";
 import { SettingsPage } from "./SettingsPage";
 import { ModalProvider } from "./context/ModalContext";
