+++ src/components/NewChatForm.tsx
@@ -222,171 +222,152 @@ export const NewChatForm: React.FC<NewChatFormProps> = ({ onClose }) => {
 
   if (showConfirmation) {
     return (
-      <div
-        className="fixed inset-0 bg-black/50 flex items-center justify-center z-[1000]"
-        onClick={handleOverlayClick}
-      >
-        <div
-          className={styles["new-chat-form"]}
-          onClick={(e) => e.stopPropagation()}
-        >
-          <h3 className={styles.title}>Confirm Handshake</h3>
-          <div className={styles["confirmation-details"]}>
-            <p>
-              <strong>Recipient:</strong> {recipientAddress}
+      <>
+        <h3 className={styles.title}>Confirm Handshake</h3>
+        <div className="text-white/80 text-sm mb-5 leading-normal">
+          <p className="my-2 break-all">
+            <strong>Recipient:</strong> {recipientAddress}
+          </p>
+          <p className="my-2">
+            <strong>Amount:</strong> {handshakeAmount} KAS
+          </p>
+          <p className="my-2">
+            <strong>Your Balance:</strong> {balance?.matureDisplay || "0"} KAS
+          </p>
+          {parseFloat(handshakeAmount) > 0.2 && (
+            <p className={styles["info-text"]}>
+              The extra amount ({(parseFloat(handshakeAmount) - 0.2).toFixed(8)}{" "}
+              KAS) helps the recipient respond even if they have no KAS.
             </p>
-            <p>
-              <strong>Amount:</strong> {handshakeAmount} KAS
-            </p>
-            <p>
-              <strong>Your Balance:</strong> {balance?.matureDisplay || "0"} KAS
-            </p>
-            {parseFloat(handshakeAmount) > 0.2 && (
-              <p className={styles["info-text"]}>
-                The extra amount (
-                {(parseFloat(handshakeAmount) - 0.2).toFixed(8)} KAS) helps the
-                recipient respond even if they have no KAS.
-              </p>
-            )}
-            {/* Only show warning if user is NOT sending extra amount */}
-            {recipientWarning && parseFloat(handshakeAmount) <= 0.2 && (
-              <p className={styles["warning-text"]}>{recipientWarning}</p>
-            )}
-            <p>This will initiate a handshake conversation. Continue?</p>
-          </div>
-          <div className={styles["form-actions"]}>
-            <button
-              type="button"
-              className={styles["cancel-button"]}
-              onClick={() => setShowConfirmation(false)}
-              disabled={isLoading}
-            >
-              Back
-            </button>
-            <button
-              type="button"
-              className={styles["submit-button"]}
-              onClick={confirmHandshake}
-              disabled={isLoading}
-            >
-              {isLoading ? "Sending..." : "Confirm & Send"}
-            </button>
-          </div>
+          )}
+          {/* Only show warning if user is NOT sending extra amount */}
+          {recipientWarning && parseFloat(handshakeAmount) <= 0.2 && (
+            <p className={styles["warning-text"]}>{recipientWarning}</p>
+          )}
+          <p>This will initiate a handshake conversation. Continue?</p>
         </div>
-      </div>
+        <div className={styles["form-actions"]}>
+          <button
+            type="button"
+            onClick={() => setShowConfirmation(false)}
+            disabled={isLoading}
+            className="w-full sm:w-fit h-12 px-5 py-2 text-sm font-bold flex items-center justify-center rounded shadow transition-all duration-200 focus:outline focus:outline-gray-300 disabled:opacity-50 disabled:cursor-not-allowed bg-gray-500 hover:bg-gray-600 active:bg-gray-700 border border-gray-500 text-white"
+          >
+            Back
+          </button>
+          <button
+            type="button"
+            onClick={confirmHandshake}
+            disabled={isLoading}
+            className="w-full sm:w-fit h-12 px-5 py-2 text-sm font-bold flex items-center justify-center rounded shadow transition-all duration-200 focus:outline focus:outline-blue-300 disabled:opacity-50 disabled:cursor-not-allowed bg-blue-500 hover:bg-blue-600 active:bg-blue-700 border border-blue-500 text-white"
+          >
+            {isLoading ? "Sending..." : "Confirm & Send"}
+          </button>
+        </div>
+      </>
     );
   }
 
   return (
-    <div
-      className="fixed inset-0 bg-black/50 flex items-center justify-center z-[1000]"
-      onClick={handleOverlayClick}
-    >
-      <div
-        className={styles["new-chat-form"]}
-        onClick={(e) => e.stopPropagation()}
-      >
-        <h3 className={styles.title}>Start New Conversation</h3>
-        <form onSubmit={handleSubmit}>
-          <div className={styles["form-group"]}>
-            <label className={styles.label} htmlFor="recipientAddress">
-              Recipient Address
-            </label>
-            <input
-              ref={useRecipientAddressRef}
-              className={styles.input}
-              type="text"
-              id="recipientAddress"
-              value={recipientAddress}
-              onChange={(e) => setRecipientAddress(e.target.value)}
-              placeholder="kaspa:..."
-              disabled={isLoading}
-              required
-            />
-            {isCheckingRecipient && (
-              <div className={styles["checking-text"]}>
-                Checking recipient balance...
-              </div>
-            )}
-            {recipientWarning && (
-              <div className={styles["warning-message"]}>
-                {recipientWarning}
-              </div>
-            )}
-          </div>
-
-          <div className={styles["form-group"]}>
-            <label className={styles.label} htmlFor="handshakeAmount">
-              Handshake Amount (KAS)
-            </label>
-            <input
-              className={styles["amount-input"]}
-              type="text"
-              id="handshakeAmount"
-              value={handshakeAmount}
-              onChange={(e) => handleAmountChange(e.target.value)}
-              placeholder="0.2"
-              disabled={isLoading}
-            />
-            <div className={styles["amount-buttons"]}>
-              <button
-                type="button"
-                className={`${styles["amount-button"]} ${
-                  handshakeAmount === "0.2" ? styles["active"] : ""
-                }`}
-                onClick={() => handleQuickAmount("0.2")}
-                disabled={isLoading}
-              >
-                0.2
-              </button>
-              <button
-                type="button"
-                className={`${styles["amount-button"]} ${
-                  handshakeAmount === "0.5" ? styles["active"] : ""
-                }`}
-                onClick={() => handleQuickAmount("0.5")}
-                disabled={isLoading}
-              >
-                0.5
-              </button>
-              <button
-                type="button"
-                className={`${styles["amount-button"]} ${
-                  handshakeAmount === "1" ? styles["active"] : ""
-                }`}
-                onClick={() => handleQuickAmount("1")}
-                disabled={isLoading}
-              >
-                1
-              </button>
+    <>
+      <h3 className={styles.title}>Start New Conversation</h3>
+      <form onSubmit={handleSubmit}>
+        <div className={styles["form-group"]}>
+          <label className={styles.label} htmlFor="recipientAddress">
+            Recipient Address
+          </label>
+          <input
+            ref={useRecipientAddressRef}
+            className={styles.input}
+            type="text"
+            id="recipientAddress"
+            value={recipientAddress}
+            onChange={(e) => setRecipientAddress(e.target.value)}
+            placeholder="kaspa:..."
+            disabled={isLoading}
+            required
+          />
+          {isCheckingRecipient && (
+            <div className={styles["checking-text"]}>
+              Checking recipient balance...
             </div>
-            <div className={styles["info-text"]}>
-              Default: 0.2 KAS. Higher amounts help recipients respond even if
-              they have no KAS. This creates a better experience for newcomers
-              to Kasia.
-            </div>
-          </div>
+          )}
+          {recipientWarning && (
+            <div className={styles["warning-message"]}>{recipientWarning}</div>
+          )}
+        </div>
 
-          {error && <div className={styles["error-message"]}>{error}</div>}
-          <div className={styles["form-actions"]}>
+        <div className={styles["form-group"]}>
+          <label className={styles.label} htmlFor="handshakeAmount">
+            Handshake Amount (KAS)
+          </label>
+          <input
+            className={styles["amount-input"]}
+            type="text"
+            id="handshakeAmount"
+            value={handshakeAmount}
+            onChange={(e) => handleAmountChange(e.target.value)}
+            placeholder="0.2"
+            disabled={isLoading}
+          />
+          <div className={styles["amount-buttons"]}>
             <button
               type="button"
-              className={styles["cancel-button"]}
-              onClick={onClose}
+              className={`${styles["amount-button"]} ${
+                handshakeAmount === "0.2" ? styles["active"] : ""
+              }`}
+              onClick={() => handleQuickAmount("0.2")}
               disabled={isLoading}
             >
-              Cancel
+              0.2
             </button>
             <button
-              type="submit"
-              className={styles["submit-button"]}
+              type="button"
+              className={`${styles["amount-button"]} ${
+                handshakeAmount === "0.5" ? styles["active"] : ""
+              }`}
+              onClick={() => handleQuickAmount("0.5")}
               disabled={isLoading}
             >
-              {isLoading ? "Initiating..." : "Start Chat"}
+              0.5
+            </button>
+            <button
+              type="button"
+              className={`${styles["amount-button"]} ${
+                handshakeAmount === "1" ? styles["active"] : ""
+              }`}
+              onClick={() => handleQuickAmount("1")}
+              disabled={isLoading}
+            >
+              1
             </button>
           </div>
-        </form>
-      </div>
-    </div>
+          <div className={styles["info-text"]}>
+            Default: 0.2 KAS. Higher amounts help recipients respond even if
+            they have no KAS. This creates a better experience for newcomers to
+            Kasia.
+          </div>
+        </div>
+
+        {error && <div className={styles["error-message"]}>{error}</div>}
+        <div className={styles["form-actions"]}>
+          <button
+            type="button"
+            onClick={onClose}
+            disabled={isLoading}
+            className="w-full sm:w-fit h-12 px-5 py-2 text-sm font-bold flex items-center justify-center rounded shadow transition-all duration-200 focus:outline focus:outline-gray-300 disabled:opacity-50 disabled:cursor-not-allowed bg-gray-500 hover:bg-gray-600 active:bg-gray-700 border border-gray-500 text-white"
+          >
+            Cancel
+          </button>
+          <button
+            type="submit"
+            disabled={isLoading}
+            className="w-full sm:w-fit h-12 px-5 py-2 text-sm font-bold flex items-center justify-center rounded shadow transition-all duration-200 focus:outline focus:outline-gray-300 disabled:opacity-50 disabled:cursor-not-allowed bg-[#2196f3] hover:bg-[#1976d2] text-white"
+          >
+            {isLoading ? "Initiating..." : "Start Chat"}
+          </button>
+        </div>
+      </form>
+    </>
   );
 };
